REFACTOR :: Architecture
